test: share sequential SQLite D1 adapter
